fix test errors
